Add option to migrate database to Lookout.
