+++ cmd/lookout/main.go
@@ -14,13 +14,17 @@ import (
 	"github.com/G-Research/armada/internal/common/serve"
 	"github.com/G-Research/armada/internal/lookout"
 	"github.com/G-Research/armada/internal/lookout/configuration"
+	"github.com/G-Research/armada/internal/lookout/postgres"
+	"github.com/G-Research/armada/internal/lookout/repository/schema"
 	lookoutApi "github.com/G-Research/armada/pkg/api/lookout"
 )
 
 const CustomConfigLocation string = "config"
+const MigrateDatabase string = "migrateDatabase"
 
 func init() {
 	pflag.String(CustomConfigLocation, "", "Fully qualified path to application configuration file")
+	pflag.Bool(MigrateDatabase, false, "Migrate databse instead of running server")
 	pflag.Parse()
 }
 
@@ -31,6 +35,18 @@ func main() {
 	var config configuration.LookoutConfiguration
 	userSpecifiedConfig := viper.GetString(CustomConfigLocation)
 	common.LoadConfig(&config, "./config/lookout", userSpecifiedConfig)
+
+	if viper.GetBool(MigrateDatabase) {
+		db, err := postgres.Open(config.PostgresConnection)
+		if err != nil {
+			panic(err)
+		}
+		err = schema.UpdateDatabase(db)
+		if err != nil {
+			panic(err)
+		}
+		os.Exit(0)
+	}
 
 	shutdownChannel := make(chan os.Signal, 1)
 	signal.Notify(shutdownChannel, syscall.SIGINT, syscall.SIGTERM)
